requirements.txt
about-time==4.2.1
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
alive-progress==3.1.5
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.4
async-timeout==4.0.3
attrs==23.2.0
autograd==1.6.2
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cma==3.2.2
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
datasets==2.20.0
dbus-python==1.3.2
Deprecated==1.2.14
diffusers==0.29.2
dill==0.3.8
distlib==0.3.8
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
gradio==4.39.0
gradio_client==1.1.1
grapheme==0.6.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.1
humanfriendly==10.0
identify==2.6.0
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
isort==5.13.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jstyleson==0.0.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
natsort==8.4.0
networkx==3.3
ninja==1.11.1.1
nncf==2.11.0
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
onnx==1.16.1
openvino==2024.4.0.dev20240724
openvino-optimum==0.0.2
openvino-telemetry==2024.1.0
openvino-tokenizers==2024.4.0.0.dev20240724
optimum @ git+https://github.com/huggingface/optimum.git@5ece6e88156de6a0de848ad6c1af381c81cfbcc3
optimum-intel @ git+https://github.com/huggingface/optimum-intel.git@819c51394216eedf9245f91f1a9db897486fff1c
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pre-commit==3.7.1
protobuf==5.27.2
psutil==6.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycairo==1.26.1
pydantic==2.8.2
pydantic_core==2.20.1
pydot==3.0.1
pydub==0.25.1
Pygments==2.18.0
PyGObject==3.48.2
pylint==3.2.6
pymoo==0.6.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
ruff==0.5.4
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
torch==2.3.1
tqdm==4.66.4
transformers==4.42.4
triton==2.3.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
virtualenv==20.26.3
watchfiles==0.22.0
websockets==11.0.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
